Update README.md
n order for B/C to have the same rules and use the nsg vnet from A the references & the rules need to be the same. My logic is to change rules in source port range & destination port range to 80 in Modules/Networking maintf. It can go from any any to any any and this will block it from going to any port. Most importantly I changed the subscription to my Azure account so the resources would deploy into the portal. The security rule in Modules/Networking is name "test03" & priority 104 on port 80. The security rules in Sub A changes to name: test1 priority:100 & example1, example2 is name:test12 priority:102 name:, example 3 "test1234" priority: 106. The "*" denotes the source & destination port range from Modules/Networking.
I check each subscription to ensure it deploys with commands of Terraform init, plan & apply in the Terminal in VSC. All subscriptions & resources deployed into the Azure portal. I push the files to my Git Hub with  GIT Status, helping to check if any changes has been made, and when identified, 
changes will be added to the Staging Environment with GIT ADD . and checking GIT status again making sure its green which means ready to be committed, 
then commit with GIT Commit -m "changes" then we push to the Repo with GIT Push.